+++ packages/tracing/src/tracer.ts
@@ -5,8 +5,7 @@ import {
   Tracer as OpenTelemetryTracer,
   TimeInput,
   SpanAttributeValue,
-  getSpan,
-  setSpan,
+  trace,
 } from "@opentelemetry/api";
 import { AsyncHooksContextManager } from "@opentelemetry/context-async-hooks";
 import {
@@ -14,11 +13,13 @@ import {
   CollectorExporterNodeConfigBase,
 } from "@opentelemetry/exporter-collector";
 import { NodeTracerProvider } from "@opentelemetry/node";
+import { Resource } from "@opentelemetry/resources";
+import { ResourceAttributes } from "@opentelemetry/semantic-conventions";
 import { SimpleSpanProcessor } from "@opentelemetry/tracing";
 
 import type { TracingConfig } from "./config";
 
-const provider = new NodeTracerProvider();
+let provider: NodeTracerProvider;
 
 let isTracerStarted = false;
 
@@ -27,12 +28,17 @@ function startTracer(options: TracingConfig, logger?: Logger): void {
     return;
   }
 
+  provider = new NodeTracerProvider({
+    resource: new Resource({
+      [ResourceAttributes.SERVICE_NAME]: getServiceName(options),
+    }),
+  });
+
   if (options.collectors) {
     for (const collector of options.collectors) {
       if (collector.type === "otel") {
         const collectorOptions: CollectorExporterNodeConfigBase = {
           attributes: options.attributes,
-          serviceName: collector.serviceName,
           url: collector.url,
         };
         if (collector.authorizationHeader) {
@@ -48,10 +54,10 @@ function startTracer(options: TracingConfig, logger?: Logger): void {
       }
     }
   }
+
   if (options.simpleCollector) {
     const collector = new CollectorTraceExporter({
       attributes: options.attributes,
-      serviceName: options.simpleCollector.serviceName,
       url: options.simpleCollector.url,
     });
 
@@ -62,7 +68,6 @@ function startTracer(options: TracingConfig, logger?: Logger): void {
   if (options.lightstepPublicSatelliteCollector) {
     const collector = new CollectorTraceExporter({
       attributes: options.attributes,
-      serviceName: options.lightstepPublicSatelliteCollector.serviceName,
       headers: {
         "Lightstep-Access-Token":
           options.lightstepPublicSatelliteCollector.accessToken,
@@ -122,7 +127,7 @@ class TracerImpl implements Tracer {
 
   withSpan<T>(name: string, callback: SpanCallback<T>): Promise<T> {
     const span = this.tracer.startSpan(name);
-    return context.with(setSpan(context.active(), span), async () => {
+    return context.with(trace.setSpan(context.active(), span), async () => {
       try {
         const result = await callback(spanFactory(span));
         span.end();
@@ -135,7 +140,7 @@ class TracerImpl implements Tracer {
   }
 
   getCurrentSpan(): Span | undefined {
-    const span = getSpan(context.active());
+    const span = trace.getSpan(context.active());
     if (span) {
       return spanFactory(span);
     }
@@ -178,7 +183,7 @@ function spanFactory(otSpan: OpenTelemetrySpan): Span {
     return this;
   };
 
-  const getTraceId = (): string => otSpan.context().traceId;
+  const getTraceId = (): string => otSpan.spanContext().traceId;
 
   const end = otSpan.end.bind(otSpan);
 
@@ -195,6 +200,18 @@ interface Span {
   setAttribute(key: string, value: unknown): this;
   setDisclosedAttribute(key: string, value: unknown): this;
   end(endTime?: TimeInput): void;
+}
+
+function getServiceName(options: TracingConfig): string {
+  if (options.collectors) {
+    return options.collectors[0].serviceName;
+  } else if (options.simpleCollector) {
+    return options.simpleCollector.serviceName;
+  } else if (options.lightstepPublicSatelliteCollector) {
+    return options.lightstepPublicSatelliteCollector.serviceName;
+  } else {
+    return "default_service_name";
+  }
 }
 
 export { getTracer, startTracer, Span, Tracer };